Major revision
Add one item min_act_to_transc to genotype. Elements in this array determine the minimal number of activators required for the transcription of a gene, i.e. the regulation logic of the gene (1 for OR GATE, 2 for AND GATE). Right now, this logic is not mutable.

A more flexible and more complicate way to implement regulation logic is to change the rate of acetylation and PIC assembly according to the number of activators and repressors. This may be explored in later revisions.