Major revision
Add function set_binding_site and argument SET_BS_MANUALLY. The function is very primitive, but allows manually configuring the topology of the network. Enable SET_BS_MANUALLY allows only mutations to kinetic constants. Note that Kon and Koff are not subjected to mutations, which means how fast Pact can be saturated by the number of a TF is currently unchangable. This will be revised in the next revision.