Major change:
The goal of this revision is to completely remove singal B and selection protein B. Essentially this is assuming that evolution has solved the problem of recognizing signal B and maximizing the growth in the presence of signal B alone. The new challenge is to correctly handle two signals, which is the project. I think this revision dramatically simplified the problem that was given to the previous model.

The revision keeps the basic structures of cisreg_cluster, which_protein, protein_pool, and so on: the first gene and the first protein always correspond to the environmental signal; the last gene and the last protein always correspond to the selection protein. However, the detail of the revision cannot be listed here.